requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.4
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2.1
django-ckeditor==6.7.2
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-js-asset==3.1.2
django-robots==6.1
djangorestframework==3.16.0
environ==1.0
idna==3.10
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
